refactor(ingredient): extract pre-parse normalization into normalize.rs
Pure move, no behavior change. Lifts the 7 pre-parse rewrite functions
(footnote/cross-ref/dimension-range/determiner/minus-equivalence strips,
normalize_input, strip_optional_note) plus the shared collapse_whitespace
out of the 949-line pipeline.rs into a focused normalize.rs. pipeline.rs
imports them via super::normalize. First step of the phase-module split.
